Stan 27.11.2019-> trzeba stworzyc pliki konfiguracyjne i eliminowac bledy

+++ Tools/TableEntries/FIB_Entry.cs
@@ -5,9 +5,9 @@ namespace Tools.Table_Entries
     public class FIB_Entry
     {
         public IPAddress destinationIP { get; set; }
-        public int portOut { get; set; }
+        public ushort portOut { get; set; }
 
-        public FIB_Entry(IPAddress destinationIp, int portOut)
+        public FIB_Entry(IPAddress destinationIp, ushort portOut)
         {
             destinationIP = destinationIp;
             this.portOut = portOut;
